requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pypdf>=4.2.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3
lxml>=5.2.2
pandas>=2.2.2
openpyxl>=3.1.5
python-pptx>=0.6.23
